Former-commit-id: 24d7e450fd1fa08817fc0bf98ea6d7b5a7df6658

+++ src/Tests/WB.Core.SharedKernels.DataCollection.Tests/InterviewTests/when_answering_real_question_with_allowed_3_decimal_places_but_answer_has_4_decimal_places.cs
@@ -51,7 +51,7 @@ namespace WB.Core.SharedKernels.DataCollection.Tests.InterviewTests
         };
 
         Because of = () => expectedException = Catch.Exception(() =>
-            interview.AnswerNumericRealQuestion(userId, realQuestionId, new int[] { }, DateTime.Now, (decimal) 0.1234));
+            interview.AnswerNumericRealQuestion(userId, realQuestionId, new decimal[] { }, DateTime.Now, (decimal)0.1234));
 
         It should_raise_AnswerDeclaredValid_event_with_QuestionId_equal_to_mandatoryQuestionId = () =>
             expectedException.ShouldBeOfType(typeof (InterviewException));
